refactor(api): split handler into modules and update documentation
- Create new modules for different handler types
- Move source collections, favorites, and submissions to separate files
- Update ApiDoc to include new paths and handlers
- Remove redundant code and improve code organization
