+++ crates/bili_sync/src/api/handlers/mod.rs
@@ -1,0 +1,5 @@
+pub mod video_handler;
+pub mod source_collections_handler;
+pub mod source_favorites_handler;
+pub mod source_submissions_handler;
+pub mod source_watch_later_handler;
